refactor: move tailwind utils to utils folder
