+++ project/components/layouts/dashboard/dashboard-content.tsx
@@ -1,4 +1,4 @@
-import { cn } from "@/lib/utils";
+import { cn } from "@/lib/utils/tailwind";
 
 interface DashboardContentProps {
   title?: string;
