+++ project/components/sections/settings/appearance-settings-form.tsx
@@ -1,6 +1,6 @@
 import Select from "@/components/ui/input-fields/select/select";
 import { Theme, useTheme } from "@/components/ui/theme-provider";
-import { cn } from "@/lib/utils";
+import { cn } from "@/lib/utils/tailwind";
 
 interface AppearanceSettingsFormProps {
   className?: string;
