+++ project/components/sections/settings/notifications-settings-form.tsx
@@ -1,6 +1,6 @@
 import Button from "@/components/ui/buttons/button";
 import Select from "@/components/ui/input-fields/select/select";
-import { cn } from "@/lib/utils";
+import { cn } from "@/lib/utils/tailwind";
 
 interface ProfileSettingsFormProps {
   className?: string;
