+++ project/components/ui/analytics-card.tsx
@@ -1,4 +1,4 @@
-import { cn } from "@/lib/utils";
+import { cn } from "@/lib/utils/tailwind";
 
 interface AnalyticsCardProps {
   title: string;
