+++ project/components/ui/buttons/link-button.tsx
@@ -1,4 +1,4 @@
-import { cn } from "@/lib/utils";
+import { cn } from "@/lib/utils/tailwind";
 import Link from "next/link";
 
 interface ButtonProps {
