+++ project/components/ui/deadline-bar.tsx
@@ -1,4 +1,4 @@
-import { cn } from "@/lib/utils";
+import { cn } from "@/lib/utils/tailwind";
 
 interface DeadlineBarProps {
   title: string;
