+++ project/components/ui/status-card.tsx
@@ -1,4 +1,4 @@
-import { cn } from "@/lib/utils";
+import { cn } from "@/lib/utils/tailwind";
 import React from "react";
 
 interface StatusCardProps {
