--- project/lib/validations.ts
@@ -1,41 +1,0 @@
-// TODO: Task 3.6 - Set up data validation with Zod schemas
-
-/*
-TODO: Implementation Notes for Interns:
-
-1. Install Zod: pnpm add zod
-2. Create validation schemas for all forms and API endpoints
-3. Add proper error messages
-4. Set up client and server-side validation
-
-Example schemas needed:
-- Project creation/update
-- Task creation/update
-- User profile update
-- List/column management
-- Comment creation
-
-Example structure:
-import { z } from 'zod'
-
-export const projectSchema = z.object({
-  name: z.string().min(1, 'Name is required').max(100, 'Name too long'),
-  description: z.string().max(500, 'Description too long').optional(),
-  dueDate: z.date().min(new Date(), 'Due date must be in future').optional(),
-})
-
-export const taskSchema = z.object({
-  title: z.string().min(1, 'Title is required').max(200, 'Title too long'),
-  description: z.string().max(1000, 'Description too long').optional(),
-  priority: z.enum(['low', 'medium', 'high']),
-  dueDate: z.date().optional(),
-  assigneeId: z.string().uuid().optional(),
-})
-*/
-
-// Placeholder exports to prevent import errors
-export const projectSchema = "TODO: Implement project validation schema";
-export const taskSchema = "TODO: Implement task validation schema";
-export const userSchema = "TODO: Implement user validation schema";
-export const listSchema = "TODO: Implement list validation schema";
-export const commentSchema = "TODO: Implement comment validation schema";
